+++ src/deconvolution.cpp
@@ -1,6 +1,6 @@
 #include "deconvolution.h"
 #include <complex>
-Deconvolution::Deconvolution(int size)
+Deconvolution::Deconvolution(unsigned int size)
 {
     _size = size;
     _pointer = 0;
@@ -12,13 +12,13 @@ Deconvolution::Deconvolution(int size)
     _inc  = new complex[_size];
     _outc = new complex[_size];
     _dc   = new complex[_size];
-    for (int i = 0; i < _size; i++) {
+    for (unsigned int i = 0; i < _size; i++) {
         _in[i] = _out[i] = 0.0;
     }
 
-    wlen = new complex[(int)std::log2(_size)];
-    float ang, dPi = 2 * M_PI;
-    for (int len = 2, l = 0; len <= _size; len <<= 1, l++) {
+    wlen = new complex[static_cast<unsigned int>(std::log2(_size))];
+    float ang, dPi = 2 * static_cast<float>(M_PI);
+    for (unsigned int len = 2, l = 0; len <= _size; len <<= 1, l++) {
         ang = dPi / len;
         wlen[l] = complex(cosf(ang), sinf(ang));
     }
@@ -44,14 +44,14 @@ void Deconvolution::add(float in, float out)
 }
 void Deconvolution::transform()
 {
-    for (int i = 0; i< _size; i++) {
+    for (unsigned int i = 0; i< _size; i++) {
         _inc[i]  = _in[i];
         _outc[i] = _out[i];
     }
 
     //direct
-    for (int i = 1, j = 0; i < _size; ++i) {
-        int bit = _size >> 1;
+    for (unsigned int i = 1, j = 0; i < _size; ++i) {
+        unsigned int bit = _size >> 1;
         for (; j>= bit; bit >>= 1)
             j -= bit;
         j += bit;
@@ -62,10 +62,10 @@ void Deconvolution::transform()
     }
 
     complex w, iu, iv, ou, ov;
-    for (int len = 2, l = 0; len <= _size; len <<= 1, l++) {
-        for (int i = 0; i < _size; i += len) {
+    for (unsigned int len = 2, l = 0; len <= _size; len <<= 1, l++) {
+        for (unsigned int i = 0; i < _size; i += len) {
             w = 1.0;
-            for (int j = 0; j < len / 2; ++j) {
+            for (unsigned int j = 0; j < len / 2; ++j) {
                 iu                     = _inc[i + j];
                 iv                     = _inc[i + j + len / 2] * w;
                 _inc[i + j]            = iu + iv;
@@ -82,13 +82,13 @@ void Deconvolution::transform()
     }
 
     //devision
-    for (int i = 0; i < _size; i++) {
+    for (unsigned int i = 0; i < _size; i++) {
         _dc[i] =  _outc[i] / _inc[i];
     }
 
     //reverse
-    for (int i = 1, j = 0; i < _size; ++i) {
-        int bit = _size >> 1;
+    for (unsigned int i = 1, j = 0; i < _size; ++i) {
+        unsigned int bit = _size >> 1;
         for (; j>= bit; bit >>= 1)
             j -= bit;
         j += bit;
@@ -98,10 +98,10 @@ void Deconvolution::transform()
     }
 
     complex u, v;
-    for (int len = 2, l = 0; len <= _size; len <<= 1, l++) {
-        for (int i = 0; i < _size; i += len) {
+    for (unsigned int len = 2, l = 0; len <= _size; len <<= 1, l++) {
+        for (unsigned int i = 0; i < _size; i += len) {
             w = 1;
-            for (int j = 0; j < len / 2; ++j) {
+            for (unsigned int j = 0; j < len / 2; ++j) {
                 u                    = _dc[i + j];
                 v                    = _dc[i + j + len / 2] * w;
                 _dc[i + j]           = u + v;
@@ -112,7 +112,7 @@ void Deconvolution::transform()
         }
     }
 
-    for (int i = 0; i < _size; i++) {
+    for (unsigned int i = 0; i < _size; i++) {
         _d[i] = _dc[i].real / _size;
     }
 }
